requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
apache-iotdb=1.3.3=pypi_0
arrow-cpp=8.0.0=py310h3098874_0
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
blas=1.0=mkl
boost-cpp=1.73.0=h7f8727e_12
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2023.05.30=h06a4308_0
certifi=2023.7.22=py310h06a4308_0
configparser=5.0.2=pyhd3eb1b0_0
gflags=2.2.2=he6710b0_0
glog=0.5.0=h2531618_0
grpc-cpp=1.46.1=h33aed49_0
icu=58.2=he6710b0_3
intel-openmp=2021.4.0=h06a4308_3561
krb5=1.19.4=h568e23c_0
ld_impl_linux-64=2.38=h1181459_1
libboost=1.73.0=h28710b8_12
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcurl=7.84.0=h91b91d3_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.46.0=hce63b2e_0
libprotobuf=3.20.1=h4ff587b_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libthrift=0.15.0=hcc01f38_0
libuuid=1.0.3=h7f8727e_2
lz4-c=1.9.4=h6a678d5_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
ncurses=6.3=h7f8727e_2
numexpr=2.8.4=py310h8879344_0
numpy=1.23.5=py310hd5efca6_0
numpy-base=1.23.5=py310h8e6c178_0
openssl=1.1.1v=h7f8727e_0
orc=1.7.4=h07ed6aa_0
packaging=22.0=py310h06a4308_0
pandas=1.5.2=py310h1128e8f_0
pip=21.2.4=py310h06a4308_0
py4j=0.10.9.3=py310h06a4308_0
pyarrow=8.0.0=py310h468efa6_0
pyspark=3.2.1=py310h06a4308_0
python=3.10.4=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2022.7=py310h06a4308_0
re2=2022.04.01=h295c915_0
readline=8.1.2=h7f8727e_1
setuptools=61.2.0=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlalchemy=1.4.54=pypi_0
sqlalchemy-utils=0.41.2=pypi_0
sqlite=3.38.3=hc218d9a_0
thrift=0.21.0=pypi_0
tk=8.6.12=h1ccaba5_0
tzdata=2022a=hda174b7_0
utf8proc=2.6.1=h27cfd23_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
